Update customizations.md
Since the figure appears correct, I decided to add figures and make any necessary wording changes to this page. So it will require a review beyond such figure positioning. As I update MUSTs on subsequent pages, I will add in the figures as needed as well. I've also gone through to fix capitalizations in line with the rules used for the Standard.

Question...there are several mentions of the map file on this page  (under fleet management and