A few more corrections and a big comment
In the second paragraph of "Back-up and Garbage Collection" It states that a copy of all files on the Image repo and its cryptographic hash should be stored on an offline system because it "allows administrators to detect and recover from a repository compromise." Should this not be explained? Or if its explained in the Standard, perhaps we should link to that?

This page really needs to be reviewed with a critical eye. When I transferred the text from the Google Doc, the original organizational structure had not been maintained. This particular page has two sections for which I could not directly match any existing text. We really need a volunteer to add text. I am going to list this lack of information as an issue.