Updates on Security page
Ok. For now, this completes my pass on the Security page. I incorporated the suggestions of @patrickvacek , added a graphic and caption and a few cross-links.
I have two content questions. First, I would propose we move the last section  "Balancing security and bandwidth" to either the Repositories section (as these sorts of decisions should be made when Uptane is first deployed) or in the Normal Operations section. I don't think it belongs here unless, perhaps, it is moved to the top of the section.

Second , under the heading  of checking dependencies, reference is made to "an NP-hard problem." I don't recall seeing this terminology anywhere else, and feel it needs to be defined. as it is used throughout this subsection.