Minor changes to exceptional ops
Can we get some eyes on the section about aftermarket ECUs? When I was reviewing this for tone, I found the last sentence in that paragraph was incomplete or just missing a few critical word or two. I attempted to fix it, but I don't know if more needed to be done.

Lois