Some edits on key management and  some questions
Can we get someone to go read through this relatively short page? The second paragraph in the online vs. offline section seems a bit contradictory. It first says "Similarly, on the Image repository, the Timestamp and Snapshot metadata SHOULD be signed using online keys so fresh metadata can be instantly generated by an automated process." But,  then the second sentence says that Offline keys can be used. I did edit these sentences to eliminate some repetition of words and phrases, but this contradiction was in the original wording as well--phrased differently but same meaning.