Correcting inconsistencies
Proposed consistent titles for both the pull down menu and the deployment page titles. As it is now, there are number of inconsistencies on what the page is called in the pull down menu as opposed to what the user sees when the page opens. I see no reason why these contents pieces shouldn't be titled the same way in both places.

Please suggest any necessary tweaks, and then I will edit the pages accordingly.
Thanks.