Responding to Issue #48
I wrote a response to Issue #48 for this page. @JustinCappos offered to do a write-up on the other two questions on the page.
The text above is basically taken from the text @JustinCappos and @tkfu already had shared on the discussion thread. I blended those comments and then @patrickvacek looked over my draft and made some suggestions.