Update exceptional_operations.md
I changed references from "pages" or "documents" to "section" to reflect the move to a hard copy. Also it is a proper name of sorts, so it needs to be capitalized.

Is "rollback" one word or two? I believe it's one word if you are describing the attack, but if you are describing the need to roll back a file, it should be two, correct?