small edits to repositories
Note that the new Table of Contents for the Deployment Best Practices features modified titles for some of the sections. The web page titles were somewhat abbreviated. The longer titles are a clearer representation of the contents and are more in line with subheads for a document.  I have been changing these titles in the in-text references.