Changing the changelog
I used the link @plapczyn provided in Issue #89 to edit and reformat the Changelog. This now should fit the stipulations in that document. The one thing I was unsure about was how in-depth  the "Unreleased" section should be. The document said it should list issues/changes not yet released. We current have over 20 such issues, including some we know may not be addressed promptly. Perhaps @plapczyn and.or @iramcdonald can provide input and advice. @tkfu this revised changelog would need to be included in your proposed Deployment Best Practices document.