Modified glossary based on @pattivacek comments
I agreed with the comments @pattivacek  made on an earlier draft of this PR (which I accidentally merged directly), so I opened this PR to address them.

She raised a much more fundamental question, that I think should be addressed to @iramcdonald. What appears here is the same glossary as in the Standard. Is that what you had in mind when you suggested the need for such a section in Deployment Best Practices? Or were you thinking we needed to evolve new terminology for this document? I would appreciate your thoughts, as well as those of others on the team.