Add text to regulations and standards
This is an extremely rough draft. I had a surprisingly difficult time coming up with short descriptions of these regulations. The official pages for the two ISO standards literally list only what iteration they were on and title.  I could find nothing in the way of a general explanation of ISO21434. I also know that there are probably a few more regulations that should be on this initial post. I just wanted to get something on the page so we can begin to get this built. So, please, add, subtract, change or point me in the right direction for better sources. Thanks.