Added information for V 1.2.0
This includes the few changes made to the Standard.

I can't remember if the initial changelog registered changes made to the Deployment Best Practices document. A huge number of changes were to those pages over the past six months. Can somebody weigh in on whether these changes need to be added as well?