Some changes and a question
Under atomic installation the second sentence reads "Uptane does not provide a way to guarantee atomic installations because the problem is out of its scope." What "problem" is this referring to? It seems to me that this is just a security option, but no problem has been defined.

Some of these corrections may be a bit nit-picky, which is why I'm doing this entire review as a pull request.