Response to Issue #86 and #87
I've added some text here in response to Issues #86 and #87. Per a decision made at the 8/3 Standards Meeting, we decided to add a piece that addresses both types of  identifiers. I based the copy on what @iramcdonald had provided in the thread as well as some of the references that text decided.

I believe once this text is approved we can close out the two issues referenced above. I am guessing that it might be awhile longer before we would want to mandate any changes on this system and likely when we do it would mean changes to the Standard. Hence, any further action would require a different type of Issue statement.