Update customizations.md
I ended up editing the existing text about transport security pretty extensively because the tone and structure of the existing copy seemed to suggest using transport security should be the exception, not the rule. would appreciate multiple reviews. @iramcdonlad, please look as well, as I was guided by your comments in doing this.