Addressing supply chain security-- Standard Issue #222 (#121)
* WIP--Adding text about supply chain security

This is still being written, so DO NOT EDIT OR MERGE at this point.

* supply chain text

Ok. I've submitted some draft text here to discuss supply chain security. It's very rough and I would welcome edits, comments, etc.

* small corrections

* Update security_considerations.md

I'm a little uncertain how to phrase what we are doing with the in-toto/Uptane mash-up as I don't want to commit us to an integration at this early stage. Hence, I re-edited this again to make the next steps a bit less certain.

* Update security_considerations.md

Co-authored-by: Marina Moore <mnm678@users.noreply.github.com>

Co-authored-by: Marina Moore <mnm678@users.noreply.github.com>