Addressing Standard Issue #221 (#120)
* Addressing Standard Issue #221

This answers Issue #221 by adding a definition of Defense-in-Depth and adding some text to the Deployment pages.

We already had a discussion of this issue in the Deployment pages, but even though it mentions "Defense-in-Depth" it perhaps could use a few more words. 

Do we need to include a  statement along the lines of "It is recommended that Uptane be used in tandem with a transport layer security system to create defense-in-depth for the vehicle" with a link to the write-up in the Deployment pages.

I'm beginning to think that the things people suggest we are lacking are just not visible because they are mentioned only in Deployment Best Practices.

* Update customizations.md

I ended up editing the existing text about transport security pretty extensively because the tone and structure of the existing copy seemed to suggest using transport security should be the exception, not the rule. would appreciate multiple reviews. @iramcdonlad, please look as well, as I was guided by your comments in doing this.

* fixed typo