Updated Changelog for V.2.0.0
This records the changes made to the Standard since V.1.2.0. The list was really not as long as I anticipated, but given that most of changes/additions were on the Deployment page (and we had agreed from the first changelog that it applied only to the Standard), it probably is the right length. Please review and let me know if there is anything I missed.