Updated Changelog for V.2.0.0 (#127)
* Updated Changelog for V.2.0.0

This records the changes made to the Standard since V.1.2.0. The list was really not as long as I anticipated, but given that most of changes/additions were on the Deployment page (and we had agreed from the first changelog that it applied only to the Standard), it probably is the right length. Please review and let me know if there is anything I missed.

* Update changelog.md

Co-authored-by: Patti Vacek <pattivacek@gmail.com>

* Update changelog.md

Co-authored-by: Patti Vacek <pattivacek@gmail.com>

* fixed a repetition

Adopted @pattivacek suggestion to remove a bullet point that repeats the idea above.

Co-authored-by: Patti Vacek <pattivacek@gmail.com>